tree & some fix

+++ gui/src/Gui_fill_map.cpp
@@ -92,7 +92,7 @@ bool Gui::fill_map(std::string data)
             std::string y = values.substr(0, values.find(" "));
             values = values.substr(values.find(" ") + 1);
             std::string orientation = values.substr(0, values.find(" "));
-            for (int i = 0; i < _players.size(); i++)
+            for (size_t i = 0; i < _players.size(); i++)
             {
                 if (_players[i].id == std::stoi(id))
                 {
@@ -124,7 +124,7 @@ bool Gui::fill_map(std::string data)
             std::string id = values.substr(0, values.find(" "));
             values = values.substr(values.find(" ") + 1);
             std::string resource = values.substr(0, values.find(" "));
-            for (int i = 0; i < _players.size(); i++)
+            for (size_t i = 0; i < _players.size(); i++)
             {
                 if (_players[i].id == std::stoi(id))
                 {
